Adding Brett Terpstra's sizes script
Terpstra created a script, called sizes, that reports the size of files
in a directory. It includes hidden files. The file size calculation is
much more accurate than that used in my dusort script. However, sizes
is a ruby script, so a ruby installation is required.

The output is color coded, which is pretty cool.

The gist for this code is here:
https://gist.github.com/ttscoff/bc6ba3cf1fffcdb11e228c12e994fed7

The blog post describing the script is here:
https://brettterpstra.com/2019/04/15/sizes-better-disk-usage-reporting-in-terminal/

	new file:   sizes