Adding the mtwrf_dircopy_nodata script
This script makes a copy of the mtWRF run directory without copying
NetCDF files and other auxiliary files. This is a useful step if you
are running another simulation with the same code by a different set
of parameters. With this script, you can quickly set up a new
directory that matches the old one.

You do not need to organize the data to get this to work.

	new file:   mtwrf_dircopy_nodata